+++ CHANGELOG.md
@@ -3,3 +3,13 @@
 ## [1.0.0] - 2024-08-22
 
 - Initial release
+
+
+## [1.0.1] - 2024-03-18
+
+- Use the new API base URL structure to call Figma’s OAuth APIs: https://api.figma.com/v1/
+
+
+## [1.0.2] - 2024-03-18
+
+- Send OAuth credentials in the request header